+++ src/main/java/com/ecommerce/ecommercewebsite/dto/ProductResponseDTO.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
@@ -15,7 +17,7 @@ public class ProductResponseDTO {
     private String productDescription;
     private ProductStatus status;
     // min price for variants and normal price for    no variants
-    private Double productPrice;
+    private BigDecimal productPrice;
     private String productCategory;
     private String productImageBase64;
     private String districtName;
